+++ html/js/dbg.js
@@ -1,0 +1,1 @@
+rtshow = ["dbg"];