+++ servers/defaults.go
@@ -94,3 +94,8 @@ var errormngt = [3]int{1, 5, 15}     // [min penalty, max panelty, max number of
 var tcpTokens chan bool              // token for accepting a TCP erquest
 var Kswitch bool                     // kill switch flag
 var RepCon bool                      // enables reporting on current
+
+// debug access control
+//var dbgMutex = &sync.Mutex{}     // lock to dbgRegistry
+//var dbgRegistry map[string]int64 // registry of curretly authorised IPs
+//const authInterval = 1           // autorisation interval for debug access in minutes